Introduce OapCacheLocations
